refactor with solid principles
