+++ ioserver/utils/liveStream.js
@@ -1,0 +1,42 @@
+import { spawn } from 'child_process';
+
+const liveStream = (streamInfo) => {
+    const options = [
+        '-i',
+        '-',
+        '-c:v', 'libx264',
+        '-preset', 'ultrafast',
+        '-tune', 'zerolatency',
+        '-r', `${25}`,
+        '-g', `${25 * 2}`,
+        '-keyint_min', 25,
+        '-crf', '25',
+        '-pix_fmt', 'yuv420p',
+        '-sc_threshold', '0',
+        '-profile:v', 'main',
+        '-level', '3.1',
+        '-c:a', 'aac',
+        '-b:a', '128k',
+        '-ar', 128000 / 4,
+        '-f', 'flv',
+        streamInfo.streamCode
+    ];
+
+    const ffmpegProcess = spawn('ffmpeg', options);
+
+    ffmpegProcess.stdout.on('data', (data) => {
+        console.log(`ffmpeg stdout: ${data}`);
+    });
+
+    ffmpegProcess.stderr.on('data', (data) => {
+        console.error(`ffmpeg stderr: ${data}`);
+    });
+    
+    ffmpegProcess.on('close', (code) => {
+        console.log(`ffmpeg process exited with code ${code}`);
+    });
+
+    return ffmpegProcess;
+}
+
+module.exports = liveStream;
